Bad orphan solution because...
Although the idea still somewhat solves the problem of defining generic
functions over TTG types, the orphan issue is not solved but rather
aggravated by this solution. For `Override` instances not to be orphans
they have to be defined next to where the ttg-stage-type was defined.
Additionally, the haddock rendering of the Override instances is
complicated and confusing :(